Added initial files via upload.
movies.json contains mock movie data generated from Mockaroo.  This is a json file containing a list of movie dictionaries.  moviesAPI.py is a preliminary API file that essentially follows the app.py demo in the 100 Days of Web course.  Currently, the GET and POST functions are working but the PUT and DELETE functions are not.  So, I'll need to do a git branch, code update, git commit, and git merge on this initial version the next time I work on this task.